Change to Lists
Made all lists public. Don't think this should be a security issue.
Also added a control for Civilian Fleets, as they were flooding the Non-Civ fleet list.
This should make things  little clearer on the front end.
My hope is to next take care of the Fleet Location -> Waypoint part, but I need to experiment a little to figure out how to do that, and how I *want* to do that.